requirements.txt
keras==2.7.0
matplotlib==3.5.2
numpy==1.22.3
pandas==1.4.1
python_dateutil==2.8.2
requests==2.28.0
scikit_learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
statsmodels==0.13.2
stqdm==0.0.4
streamlit==1.10.0
tqdm==4.64.0
xgboost==1.5.1
